+++ modules/laptop.nix
@@ -1,12 +1,17 @@
 {
   imports = [ "wireless" ];
 
-  nixosModule = { ... }: {
+  nixosModule = { pkgs, ... }: {
     services.xserver.libinput.enable = true;
 
     services.udev.extraHwdb = ''
       evdev:name:AT Translated Set 2 keyboard:dmi:*
         KEYBOARD_KEY_3a=esc
+    '';
+
+    # Add HandlePowerKeyLongPress when we're using systemd 250
+    services.logind.extraConfig = assert (builtins.compareVersions pkgs.systemd.version "250" == -1); ''
+      HandlePowerKey=ignore
     '';
   };